requirements.txt
selenium
webdriver_manager
pandas
python-dotenv
smtplib
twilio
BeautifulSoup